Make `stbt.Keypress` (return type of `stbt.press`) a public API
...and move it from `_stbt.core` to `_stbt.types`. This fixes a
regression in the previous commit where we'd need `gi` installed
to import stbt_core.
